requirements.txt
boto3
PyMuPDF
botocore
flask
flask_cors
fuzzywuzzy
jinja2
python-dateutil
python-Levenshtein
llama-index>=0.12.3
langchain-huggingface>=0.1.2
anthropic>=0.40.0
torch>=2.5.1
sentence-transformers>=3.3.1
python-dotenv>=1.0.1
llama-index-embeddings-langchain>=0.3.0
llama-index-llms-anthropic>=0.5.0
pypdf
python-dotenv
pdfplumber
pdfminer.six
defusedxml
langchain
langchain-community
python-docx
defusedxml
typing-extensions
requests
beautifulsoup4
flask-restx>=1.3.0

